First implementation of GroupBy (returning collections)
this implementation of GroupBy assumes that the input is already sorted on the keys used
for the grouping. the query compiler will put such a corresponding Sort operator in front
of each GroupBy operator. the GroupBy operator then just looks at each element's keys and
whenever the keys change from the current group's keys a group is returned.