Expect groups to be aggregated within tasks
don't transfer group_end markers across node boundaries as the group markers
are not handled correctly by the (round-robin) absorber when multiple channels
have to be absorbed (e.g. after shuffling or when connected tasks have different
DOPs) or when regrouping is necessary after shuffling. at the same time, when
nothing is shuffled but tasks with same DOP are connected point-to-point, we
don't see any benefit in just doing the grouping in a different task.