Operator to update a mutable dataset with the data stream [WIP]
the update operator requires an update-UDF as well as key indices and needs to be
co-located to the mutable data set that should be updated. the operator then takes
every element and when the elements keyset matches the keyset of an element in the
mutable dataset, the UDF is called. non-null return values from the UDF then are
updates to the dataset.

missing: working mutable datasets (get the dataset driver to handle mutable datasets
as well) and an interface to manipulate mutable datasets from an operator (get, put,
contains)