Store TMs in zookeeper persistently (no remove on disconnects)
zookeeper connections may timeout on the cluster and then reconnect automatically
through the Curator framework. if the TaskManager info is however stored as EPHEMERAL
data in zookeeper, the data is deleted upon connection loss even when the TaskManagers
reconnect to the zookeeper server automatically immediately.

storing the information as PERSISTENT data prevents the data from getting removed on
connection losses. however, this way the data is not removed at all: the WorkloadManager
has currently no knowledge about TaskManagers that actually have been removed.